Add "zoom" feature, Minor usability improvements
To use the area outside the markers in the output image, you can now set the zoom to <100%. This is required e.g. if there is not enough space above the laser bed, so the top edge of the top-left marker is below the top edge of the laser bed -- with the new setting, you can still see the whole laser bed; before, the area above the top-left marker was cut off.

Also make the web page look a little more friendly, add help links etc.